+++ tests/acompat.py
@@ -118,7 +118,7 @@ class AEvent(asyncio.Event):
         await asyncio.wait_for(self.wait(), timeout)
 
 
-class Queue(queue.Queue):  # type: ignore[type-arg]  # can be dropped after Python 3.8
+class Queue(queue.Queue):  # type: ignore[type-arg]
     """
     A Queue subclass with an interruptible get() method.
     """