+++ docs/news.rst
@@ -13,8 +13,8 @@ Future releases
 Psycopg 3.2.4 (unreleased)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-- Don't lose notifies received between two `~Connection.notifies()` calls
-  (:ticket:`#962`).
+- Don't lose notifies received whilst the `~Connection.notifies()` iterator
+  is not running (:ticket:`#962`).
 - Make sure that the notifies callback is called during the use of the
   `~Connection.notifies()` generator (:ticket:`#972`).