+++ docs/news.rst
@@ -13,6 +13,8 @@ Current release
 Psycopg 3.2.4
 ^^^^^^^^^^^^^
 
+- Don't lose notifies received whilst the `~Connection.notifies()` iterator
+  is not running (:ticket:`#962`).
 - Make sure that the notifies callback is called during the use of the
   `~Connection.notifies()` generator (:ticket:`#972`).